:truck: Rename type file
